localized getAttributes
